Add BaseFragment for FragmentTransaction methods, Add Back Button for Rate and Setting Fragments, Update Array String for Language Array
