+++ app/src/main/java/com/delacrixmorgan/kingscup/menu/MenuRateFragment.kt
@@ -6,6 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.delacrixmorgan.kingscup.R
+import kotlinx.android.synthetic.main.fragment_menu_rate.*
 
 /**
  * Created by Delacrix Morgan on 11/11/2017.
@@ -13,13 +14,14 @@ import com.delacrixmorgan.kingscup.R
 
 class MenuRateFragment : Fragment() {
 
-    override fun onCreateView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View {
-        val rootView = inflater!!.inflate(R.layout.fragment_menu_rate, container, false)
-
-        return rootView
-    }
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View =
+            inflater.inflate(R.layout.fragment_menu_rate, container, false)
 
     override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        
+        this.backButton.setOnClickListener {
+            fragmentManager.popBackStack()
+        }
     }
 }
