+++ app/src/main/res/layout/fragment_menu_setting.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/colorAccent"
+    android:background="@color/colorPrimaryDark"
     app:corner_radius="16dp">
 
     <android.support.design.widget.FloatingActionButton
